arch: x86: initial commit of x86 instruction set
This is a very crude early implementation of the x86 instruction set for
sas. Many more mnemonics/opcodes have yet to be defined, and currently
there are no instructions that handle memeory references.

In general, any instruction that is currently implemented only handles
register arguments (with a few exceptions), but the instructions
implemented should handle most/all register arguments (except for some
of the more unique combinations, typically involving the "a" registers,
eax/ax/ah/al).