core: correct the number calculation
The number calculation was incomplete/incorrect for hex numbers encoded
as 0Nh, causing the 0Nh numbers to be caught and converted, but also
catching and converting hex numbers passed in raw (sas operates in pure
hex internally, and the default for all numbers is hex).

I had originally named the conversion functions without the convert_
prefix, and the conversions in num() were still referencing the old (and
no longer existent) names. Updated the function calls to match the
correct, prefixed names.

Added a zero padding to bytes with only single (hex) digit values,
otherwise the output later would be broken.