core: learned get_reg
Instead of having to type \$${1,2} every time we are passed operand
input that happens to string-match for a register, put the dereference
in a function so we can just do $(get_reg $1) in the code, which is far
cleaner and easier to maintain (and less prone to typo bugs).

The code that already exists in x86.set has been converted to use this
new get_reg function.