core: warn user not to touch IFS
Warn the user via source comment that IFS is *not* a user-level
variable. They can modify it to get different effects, but only if they
really know what they are doing. It is there to force a sane value so
that we don't get tripped up from inheriting some wacky value from a
calling shell.