core: optimize by avoiding sed
It turns out that sed is very CPU-intensive, and avoiding it greatly
improves the speed of the assembler. To the extent possible, sed is not
used anywhere in the core (the core is the only place where external
binaries are called). In fact, there is only a single call to sed, and
it's in the same place as the single call to readlink, which is in the
initialization of the main global variables at startup, so they only
happen once and can be removed if the value is hardcoded.

Thanks to Robert Gerus <ar@bash.org.pl> for the suggestions on avoiding
sed.