core: learned run(), run external apps indirectly
Implemented a wrapper funtion which handles the execution of external
binaries. The wrapper function ensures that the binary being run is in
fact the system binary, and not a shell builtin or override, by using
the POSIX standard command builtin. It also uses the  -p flag, which
according to the standard assures that the binary will be searched for
and executed from a sane PATH setting.

All the external binary calls in core (the core is the only place
external calls should ever be made) have been converted to pass through
the new run() function.