core: generalize and abstract string printing
Most of the printf calls are very repetitive, so having a wrapper
function not only reduces typing, it also ensures uniformity and can
reduce mistakes in generating correct output. The output() functions
provide standardized interfaces to the printf binary, and ensure that
printf (which is the most-used external binary in this project) is
always run through the run() wrapper function.

This abstraction over output also enables the dispensing of echo, which
is less standardized across platforms and shells anyway.