core: get_reg can process multiple registers
Previously get_reg only processed a single register, so if you wanted to
get the values of multiple register variables, you had to make multiple
calls to get_reg. It will now process however many arguments are passed
to it via a loop, so the number of register values that can be obtained
in a single call is only limited by memory or the constraints of the
shell. Currently there are no implemented instructions which make use of
this new functionality, but it may become useful later on.