core: tohex() learns to test total byte length
If tohex() is passed a second argument parameter, it will be used to
test the length of the converted number. This can be used to ensure that
it falls within a required range of bytes (2 hex digits = 1 byte, so if
$2 == 2 then the output should be 1 byte, $2 == 4 would be one word,
etc.). This is primarily useful for error checking in the ISA set files,
where instruction opcodes may require different sizes of input numbers.