core: pad with lead 0 in tohex(), not tobytes()
If a converted hex number yields a value that consists of an odd number
of digits, pad the result with a single leading 0. This will ensure the
value matches for number input size when tohex() is used to test
againsts the second argument (i.e., tohex() $number 2), as well as avoid
needing a second loop in tobytes(). If the lead 0 isn't wanted, it can
just be passed through strip_zero().