x86.set: comments to explain the register groups
Although the register groupings should be obvious to anyone who is
familiar with x86 assembly language, they will likely be non-obvious to
anyone who is not. Adding comments helps explain why the registers are
grouped the way they are, so that x86.set can serve as a model for later
ISA set file creators who work on other ISAs and happen to not be
familiar with x86.