core: exit on invalid
Rather than a simple "return 1", which will allow execution to continue,
commit seppuku if passed an invalid instruction. If execution continues,
then the error message could be buried and the user may think a
successful assembly took place, when in fact there was an error (or
possibly multiple). By forcing an immediate exit, it should be clear to
the user that the resulting assembled code is invalid.