Roadmap: updated to reflect project status/goals
Updated the roadmap to reflect the current project status and bring it
in sync with the timeline of the immediate goals, which has shifted
slightly. Also made it a little more verbose so it is more clear
precisely what the (short-term) goals are. Long-term goals are of course
always a bit nebulous, at least in terms of a timeline.