core: learned offset() for PC-relative offset addr
Implemented offset(), which performs "$1 - ($sas_pc + $2)". The second
argument is optional, and is used for adjusting the value of the PC by
the number of bytes in the current instruction (which is yet to be
assembled). Under x86, PC-relative instructions are always relative from
the PC address of the _next_ instruction, so for x86 the second argument
is required.