core: learned option to flip endianness
It is now possible to flip the endianness of all assembled output via
the -e option flag. This is particularly useful for ISAs the support
more than one endianness, such as MIPS. This way a single <ISA>.set file
can be implemented, defaulting to one ordering, and the alternative byte
order can be achieved via the option, reducing the amount of ISA
implementation required.