core: change number parsing for decimal/hex
There was a potential problem with the parsing of decimal numbers vs.
hexadecimal numbers. A multi-digit hexadecimal number that happens to
contain low enough numbers for each digit to fit with decimal, but had
the lowest digit as hexadecimal 'd' (decimal 13) would have been
ambiguous between a decimal number minus the lowest 'd' and the
hexadecimal number. Therefore, hexadecimal strings must now be either
prefixed with '0x' or suffixed with 'h' and bare numbers are
interpreted as decimal digit strings, in order to avoid possible
ambiguities or conflicts.