new function, check_type
Implemented a new function in the core, check_type(), which will compare
an input value against a space-separated string of patterns. This is
useful for ensuring that a given operand matches a required register
space, or for ensuring that an operand is an immediate value. One could
even construct patterns to ensure that an operand is of a given size as
well. The pattern space is expanded following the rules for case block
pattern expansion (currently the function actually uses case as a way of
forcing this expansion).

The x86.set instruction set implementation was updated to make use of
the new check_type function (and also adjusted for the change to pad for
byte vs. digit padding).