new function address() and update x86.set
Added a new function, address(), which wraps tohex and pad to produce a
byte-aligned address of a given size/length from a passed in immediate.
Technically the passed in number should already be in well-formed hex,
so I may remove the tohex call from address() later if it's really not
needed.

Some instructions in x86.set were updated to make use of the new
functions (check_type, memref, and address) in combination to provide
mnemonic->opcode mappings for direct and indirect memory references. In
particular the mov instruction is now nearly complete.

I also updated the register pattern match definitions in x86.set to make
them more accurate.