core: new function, check_size
Implemented a new function, check_size(), which is a wrapper around
address() (which already tests to make sure the result matches the
specified size). The difference between check_size and address is that
address actually outputs its result if the size matches, whereas
check_size only returns true or false depending on the result, without
actually outputting the result. This is used in x86.set, for example, to
make sure that input from the user does not exceed a possible size
before attempting to assemble it.