core: correct handling of direct assembly (-a)
When called with the '-a' flag for direct assembly, rather than taking
the full command-line, which could potentially still include other flags
as well depending on when the '-a' flag appears/is processed, take a
single argument and assemble it. This now requires that instructions
that take operands must be passed in quoted to avoid the controlling
shell from splitting on the spaces.