core: pure shell implementation of tr
The use of tr to convert commas to spaces in input code lines has been
replaced by two new pure shell functions. The functions responsible for
this are substr(), which searches an input string for a given substring,
and replace(), which uses substr() to recursively replace a given
substring with another.

I'm not sure if this is faster or slower, as I haven't done any real
benchmarking, but it does increase portability by relying on one less
external program. The external dependencies are now strictly limited to
a single instance of cat (for displaying the help message only) and
printf, which is abused as a real workhorse. In the future I will likely
make the use of printf optional, as shells with a builtin printf would
likely execute faster than relying on the external printf, though there
can be differences in the implementations for how they process their
inputs, so the external printf will always be the default.