core: don't print 'command not found' for bad ops
If the user passes in an unrecognized instruction, such as 'foo eax,ebx'
for x86, and no instr_foo() exists in x86.set, then no matching
instruction can be found. Since POSIX shell doesn't enable any way to
test for the definition of a function before calling the function, sas
will simply try to call the function but check the return code for error
afterwards. This results in the user seeing an error message directly
from the shell of 'command not found'. Redirecting the STDERR output to
/dev/null makes this much cleaner, since sas will itself report that it
is an invalid instruction anyway.