core: selectable number of bits for memory output
When verbose or debug is enabled, sas will output memory offsets for
each instruction it encodes. The memory offsets were hardcoded as
32-bit. This commit now provides the option for setting the number of
bits, so that future 64-bit ISAs will not print with 32-bit offset
addresses.

Currently this is really only useful for verbose messages and
informational viewing by the user, as it doesn't affect any actual
assembly. This is because sas is currently only capable of raw binary
output. If at a later stage other output formats, such as ELF, are
implemented as modules, then the architecture bits will become more
important.