core: reimplement check_size()
Instead of abusing address() (which itself abuses pad()), test the size
of the input string directly. This is more reliable, as it's possible,
though unlikely, that the definition of address() may need to change in
the future to test for something other than byte sizes (ie, it may test
directly for actual "address" sizes, such as 8 bytes on 32-bit
platforms). Additionally, check_size can test for non-byte-aligned sizes
this way.