core: optimize tohex() a bit
When using tohex() to test the size of a given argument, there's no need
to save the argument into a variable and perform algebra on it
separately. Thus, this step can be removed to make tohex() a tiny bit
faster and require slightly less memory.

Instead, the extra padding is added in num(), since the use of num()
requires bytes. The num() function was additionally modified to
guarantee that the input produces a valid number before trying to pad it
and convert it to bytes, in order to prevent erroneous output.