core: call programs directly when possible
The print functions are called, at a minimum, for every byte that gets
output by the assembler. It is therefore important that they be able to
process quickly. Changing from calling run() to directly calling printf
(wrapped with the 'command' program to ensure it's loaded from the
proper path) results in a noticeable speedup.